\name{AnglerCreelSurveySimulation-package}
\alias{AnglerCreelSurveySimulation-package}
\alias{AnglerCreelSurveySimulation}
\docType{package}
\title{Simulate a Bus Route Creel Survey of Anglers}
}
\description{Create an angler population, sample the population with a user-specified survey times, and calculate metrics from a bus route-type creel survey.}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AnglerCreelSurveySimulation}
\packageIndices{AnglerCreelSurveySimulation}
}
\author{Steven Ranney}

Maintainer: \packageMaintainer{AnglerCreelSurveySimulation}
}

\keyword{ package }


