% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_times.R
\name{check_times}
\alias{check_times}
\title{Check that start, wait, or end times for a surveyor are provided and reasonable}
\usage{
check_times(
  start_time = NULL,
  wait_time = NULL,
  end_time = NULL,
  fishing_day_length = NULL
)
}
\arguments{
\item{start_time}{The start time of the clerk. An \code{int} between 0 and 
\code{fishing_day_length}. The default is \code{NULL}.}

\item{wait_time}{the wait time of the clerk. The default is \code{NULL}. The default is 
\code{NULL}}

\item{end_time}{the end time of the clerk. An \code{int} between 0 and 
\code{fishing_day_length}. The default is \code{NULL}.}

\item{fishing_day_length}{the total length of the fishing day, in hours, as \code{int}. 0800hrs to 2000hrs = 
fishing_day_length of 12 hours.}

\item{circuit_time}{the total time it takes a surveyor to complete their sampling circuit.}
}
\description{
This is a helper function used in \code{get_total_value} to check that 
surveyor for \code{start_time}, \code{wait_time}, or \code{end_time} 
are valid.
}
\details{
Two of the three \code{start_time}, \code{wait_time}, or \code{end_time} must be 
provided. The third will be calculated.
}
\examples{
library(dplyr)   
set.seed(256)

start_time = NULL
end_time = NULL
wait_time = NULL

assign_generic_wait_time(start_time = start_time, end_time = end_time, wait_time = wait_time, 
fishing_day_length = 12)

start_time = 2
end_time = NULL
wait_time = NULL

assign_generic_wait_time(start_time = start_time, end_time = end_time, wait_time = wait_time, 
fishing_day_length = 8)
                   
}
\author{
Steven H. Ranney
}
